Curl correct cluster IP from test runner pod
